Add zdaemon.tests to the list of install packages in setup().
In zdaemon, print the wait status if it is different than the signal
that caused the process to be terminated.  This message will often
signal that a core file has been produced.  We don't say that directly
because there's no standard test for it.  The wait status is the next
best thing, I think, and it's what zdaemon printed until the previous
checkin.