Merge changes made from a wrong project:
r28912 | tim_one | 2005-01-21 17:22:52 -0500 (Fri, 21 Jan 2005) | 17 lines
Changed paths:
   M /Zope/trunk/lib/python/zdaemon/tests/testzdrun.py

Try to repair testRunIgnoresParentSignals.

Instead of poke-and-hope guessing at how long to wait
for the child process to start running, use send_action()
in a loop to ask about that more directly -- but don't
wait longer than a minute.

The test should run faster on most boxes now, but
take up to a minute on boxes that are bogged down for
whatever reason.

Note that this checkin is being made from a wrong place!
zdaemon isn't part of the Zope tree, it's stitched in
from the zdaemon repository.  So this checkin has no
effect on zdaemon.  It will need to be sideported later
if it turns out to fix the overnight test failures.