Track coverage in subprocesses
This currently only works for ./zdaemon executed from the README.rst
test.  Other tests need to be modified in a similar way (i.e. call
coverage.process_start() in the child process, if 'coverage' is
importable).

Warning: if you intend to run 'coverage report' or 'coverage xml', you
need to explicitly set COVERAGE_HOME=. or it will fail to find the
covergae information.