Fix delayed deadlock when transcript thread dies (#1)
Test plan:

- mount a small tmpfs with 'sudo mount -t tmpfs none /mnt -o size=4096'

- create a zdaemon config file
    <runner>
      program yes
      transcript /mnt/transcript.log
    </runner>

- run zdaemon -C conf start

- wait a few milliseconds for /mnt to fill up

- pgrep yes

If the 'yes' program is still running, we have a deadlock (strace and
you'll find it blocked on write()).  This is the situation before this
patch, as described in bug #1.

If the 'yes' program is dead, the deadlock is fixed.

- run zdaemon -C conf status

The daemon manager should be stopped (it's not functional without the
dead transcript thread).

Automating this test is left as an exercise for the reader.  :(